rename for clarity add comments and follow other implemented conventions and upcoming changes
