+++ src/components/particles-sketch/particles-sketch.css
@@ -1,0 +1,16 @@
+#background-sketch-container {
+  position: absolute;
+  display: block;
+  width: 100%;
+  height: 100%;
+  /* prevent pointer when not on screen */
+  pointer-events: none;
+  opacity: 0%;
+  transition: opacity var(--default-transition-dur) ease-in-out;
+}
+
+/* Class selector appended to show-hide elements using opacity */
+#background-sketch-container.on-screen {
+  opacity: 100%;
+  pointer-events: auto;
+}
